# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/stations": {
            "get": {
                "summary": "Get info about all stations",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Station"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create station",
                "parameters": [
                    {
                        "description": "Serial number of station",
                        "name": "SerialNumber",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Address Id",
                        "name": "Address",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Full capacity on station",
                        "name": "Capacity",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "description": "Free capacity on station",
                        "name": "FreeCapacity",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Station"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/stations/{stationId}": {
            "get": {
                "summary": "Get info about station",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Station Id",
                        "name": "stationId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Station"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "summary": "Update station",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Station Id",
                        "name": "stationId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Address Id",
                        "name": "Address",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Free capacity on station",
                        "name": "FreeCapacity",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Station"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Delete station",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Station Id",
                        "name": "stationId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/stations/{stationId}/powerbanks": {
            "get": {
                "summary": "Get powerbanks in station",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Station Id",
                        "name": "stationId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Powerbank"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Take random powerbank from station",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Station Id",
                        "name": "stationId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/stations/{stationId}/powerbanks/{powerbankId}": {
            "put": {
                "summary": "Return powerbank to station",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Station Id",
                        "name": "stationId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Powerbank Id",
                        "name": "powerbankId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Add powerbank to station",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Station Id",
                        "name": "stationId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Powerbank Id",
                        "name": "powerbankId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Powerbank's position in station",
                        "name": "Position",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "entity.Address": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lat": {
                    "type": "number"
                },
                "lng": {
                    "type": "number"
                }
            }
        },
        "entity.Powerbank": {
            "type": "object",
            "properties": {
                "capacity": {
                    "description": "объем заряда",
                    "type": "number"
                },
                "create_at": {
                    "type": "string"
                },
                "delete_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "position": {
                    "type": "integer"
                },
                "removed": {
                    "type": "boolean"
                },
                "serial_number": {
                    "type": "string"
                },
                "update_at": {
                    "type": "string"
                },
                "used": {
                    "description": "сколько уже использована банка в часах // up юзается или нет",
                    "type": "boolean"
                }
            }
        },
...
        "entity.Station": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "integer"
                },
                "address_full": {
                    "$ref": "#/definitions/entity.Address"
                },
                "capacity": {
                    "type": "number"
                },
                "create_at": {
                    "type": "string"
                },
                "delete_at": {
                    "type": "string"
                },
                "free_capacity": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "powerbanks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Powerbank"
                    }
                },
                "removed": {
                    "type": "boolean"
                },
                "serial_number": {
                    "type": "string"
                },
                "update_at": {
                    "type": "string"
                }
            }
        },
...
    }
}